feat: Move pause screen to separate camera
In order to use the world-space UI, the pause screenshot must be in camera space. This can be a problem if the scene clips through the image. The best solution, therefore, is to render the pause screen using its own camera which we assume will never clash with the game world.