feat: Add more audience models with different materials
If a blender mesh has multiple materials and you export it to .fbx, Unity will ignore all materials except the first one (not just for the mesh renderer, but also for the asset itself). This means that if you want to export "unused" materials, you have to export a mesh that uses each one. The alternative method is to create the materials in Unity, which I could certainly do because it's a basic diffuse material with a single texture.